Cambios organizacionales al repositorio:
 - Mover storeCreator a su propia carpeta.
 - Separar y organizar reducers.
 - Ajustar la aplicación para que tenga en cuenta estos detalles.
 - Agregado archivo de licencia (ISC).
 - Modificado README.MD para que sea más informativo.
