+++ README.MD
@@ -1,5 +1,28 @@
 # Frontend UTNianos
 
-## Correr el repositorio
+## Arrancar el repositorio
 
+### Instalar las dependencias y compilar el repositorio.
+```sh
+npm install
+webpack --progress
+```
 
+### Iniciar un servidor en la carpeta donde esta el JS.
+
+Cualquier servidor de archivos estáticos funciona, por ejemplo 
+http.server de python o npm.
+ 
+```sh
+cd build
+python -m http.server
+http://localhost:8000/
+```
+
+Ó
+
+```sh
+cd build
+npm start
+http://localhost:3000/
+```
