--- src/storeCreator.js
@@ -1,11 +1,0 @@
-import { createStore, combineReducers, applyMiddleware, compose } from 'redux';
-import thunkMiddleware from 'redux-thunk';
-import * as reducers from 'reducers';
-
-export default function storeCreator() {
-  const reducer = combineReducers(reducers);
-  return createStore(reducer, compose(
-      applyMiddleware(thunkMiddleware),
-      window.devToolsExtension ? window.devToolsExtension() : f => f
-  ));
-}
